+++ core/src/main/java/screens/MenuScreen.java
@@ -1,12 +1,10 @@
 package screens;
 
-import com.badlogic.gdx.audio.Sound;
 import com.badlogic.gdx.scenes.scene2d.ui.ImageButton;
 import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.principal.AntsArmageddon;
 import hud.EventosBoton;
 import hud.FabricaBotones;
-import managers.GestorAssets;
 import utils.Constantes;
 
 public class MenuScreen extends ScreenMenus {
@@ -18,11 +16,9 @@ public class MenuScreen extends ScreenMenus {
     @Override
     protected void construirUI() {
 
-        Sound sonidoClick = GestorAssets.get(Constantes.SONIDO_CLICK, Sound.class);
-
-        ImageButton jugar = FabricaBotones.JUGAR.crearBoton(sonidoClick, EventosBoton.irJuego(juego));
-        ImageButton opciones = FabricaBotones.OPCIONES.crearBoton(sonidoClick, EventosBoton.irMenuOpciones(juego));
-        ImageButton salir = FabricaBotones.SALIR.crearBoton(sonidoClick, EventosBoton.salirJuego());
+        ImageButton jugar = FabricaBotones.JUGAR.crearBoton(Constantes.ATLAS_BOTONES, Constantes.SONIDO_CLICK, EventosBoton.irPreGameScreen(juego));
+        ImageButton opciones = FabricaBotones.OPCIONES.crearBoton(Constantes.ATLAS_BOTONES, Constantes.SONIDO_CLICK, EventosBoton.irMenuOpciones(juego));
+        ImageButton salir = FabricaBotones.SALIR.crearBoton(Constantes.ATLAS_BOTONES, Constantes.SONIDO_CLICK, EventosBoton.salirJuego());
 
         Table table = new Table();
         table.setFillParent(true);