+++ core/src/main/java/screens/OpcionesScreen.java
@@ -1,12 +1,10 @@
 package screens;
 
-import com.badlogic.gdx.audio.Sound;
 import com.badlogic.gdx.scenes.scene2d.ui.ImageButton;
 import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.principal.AntsArmageddon;
 import hud.EventosBoton;
 import hud.FabricaBotones;
-import managers.GestorAssets;
 import utils.Constantes;
 
 public class OpcionesScreen extends ScreenMenus {
@@ -18,11 +16,9 @@ public class OpcionesScreen extends ScreenMenus {
     @Override
     protected void construirUI() {
 
-        Sound sonidoClick = GestorAssets.get(Constantes.SONIDO_CLICK, Sound.class);
-
-        ImageButton sonido = FabricaBotones.SONIDO.crearBoton(sonidoClick, EventosBoton.descomponerAtlas());
-        ImageButton graficos = FabricaBotones.OPC1.crearBoton(sonidoClick, EventosBoton.irPreGameScreen(juego));
-        ImageButton volver = FabricaBotones.VOLVER.crearBoton(sonidoClick, EventosBoton.salirMenuOpciones(juego));
+        ImageButton sonido = FabricaBotones.SONIDO.crearBoton(Constantes.ATLAS_BOTONES, Constantes.SONIDO_CLICK, EventosBoton.descomponerAtlas());
+        ImageButton graficos = FabricaBotones.OPC1.crearBoton(Constantes.ATLAS_BOTONES, Constantes.SONIDO_CLICK, EventosBoton.irPreGameScreen(juego));
+        ImageButton volver = FabricaBotones.VOLVER.crearBoton(Constantes.ATLAS_BOTONES, Constantes.SONIDO_CLICK, EventosBoton.salirMenuOpciones(juego));
 
         Table table = new Table();
         table.setFillParent(true);